Enum js_option::JsOption

pub enum JsOption<T> {
    Some(T),
    Null,
    Undefined,
}

An Option-like type with two data-less variants in addition to Some: Null and Undefined.

Variants

Some(T)

Some value T

Null

Explicit absence of a value

Undefined

Implicit absence of a value

Implementations

impl<T> JsOption<T>

pub fn from_option(opt: Option<T>) -> Self

Construct a JsOption from a regular Option.

None will be converted to to Null.

pub fn from_implicit_option(opt: Option<T>) -> Self

Construct a JsOption from a regular Option.

None will be converted to Undefined.

pub fn into_option(self) -> Option<T>

Convert a JsOption to Option.

pub fn into_nested_option(self) -> Option<Option<T>>

Convert a JsOption<T> to Option<Option<T>>.

Null is represented as Some(None) while Undefined is represented as None.

pub const fn is_some(&self) -> bool

Returns true if the JsOption contains a value.

pub const fn is_null(&self) -> bool

Returns true if the JsOption is Null.

pub const fn is_undefined(&self) -> bool

Returns true if the JsOption is Undefined.

pub fn unwrap(self) -> T

Returns the contained Some value, consuming the self value.

Panics

Panics if the self value equals Null or Undefined.

pub fn unwrap_or(self, default: T) -> T

Returns the contained Some value or a provided default.

pub fn unwrap_or_else<F: FnOnce() -> T>(self, f: F) -> T

Returns the contained Some value computes is from a closure.

pub fn map<U, F: FnOnce(T) -> U>(self, f: F) -> JsOption<U>

Maps a JsOption<T> to JsOption<U> by applying a function to a contained value.

pub const fn as_ref(&self) -> JsOption<&T>

Converts from &Option<T> to Option<&T>.

pub fn as_mut(&mut self) -> JsOption<&mut T>

Converts from &mut Option<T> to Option<&mut T>.

impl<T: Default> JsOption<T>

pub fn unwrap_or_default(self) -> T

Returns the contained Some value or a default.

impl<T: Deref> JsOption<T>

pub fn as_deref(&self) -> JsOption<&<T as Deref>::Target>

Converts from &JsOption<T> to JsOption<&T::Target>.

impl<T: DerefMut> JsOption<T>

pub fn as_deref_mut(&mut self) -> JsOption<&mut <T as Deref>::Target>

Converts from &mut JsOption<T> to JsOption<&mut T::Target>.